chore: make seperate file
